move fd
